Repackage FindCommand and remove Filter
