+++ src/main/java/seedu/address/logic/commands/findcommand/FindEmailCommand.java
@@ -1,16 +1,16 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.findcommand;
 
 import static java.util.Objects.requireNonNull;
 
+import seedu.address.logic.commands.CommandResult;
 import seedu.address.model.Model;
-import seedu.address.model.person.NameContainsKeywordsPredicate;
+import seedu.address.model.person.EmailContainsKeywordsPredicate;
 /**
- * Finds and lists all persons in address book whose name contains any of the argument keywords.
- * Keyword matching is case-insensitive and allows partial matching.
+ * Finds and lists all persons in address book whose email contains any of the argument keywords.
+ * Keyword matching is case-insensitive and allows partial matching, including numbers and symbols.
  */
-public class FindNameCommand extends FindCommand {
-
-    public static final String MESSAGE_FIND_NAME_PERSON_SUCCESS = "Search for name containing \"%s\" was successful. "
+public class FindEmailCommand extends FindCommand {
+    public static final String MESSAGE_FIND_EMAIL_PERSON_SUCCESS = "Search for email containing \"%s\" was successful. "
             + " Showing results:";
 
     /**
@@ -20,17 +20,17 @@ public class FindNameCommand extends FindCommand {
      *
      * @param predicate Keywords used to filter contacts by name.
      */
-    public FindNameCommand(NameContainsKeywordsPredicate predicate) {
+    public FindEmailCommand(EmailContainsKeywordsPredicate predicate) {
         super(predicate);
     }
 
     @Override
     public CommandResult execute(Model model) {
         requireNonNull(model);
-        model.updateFilteredPersonList((NameContainsKeywordsPredicate) predicate);
+        model.updateFilteredPersonList((EmailContainsKeywordsPredicate) predicate);
 
         if (!model.getFilteredPersonList().isEmpty()) {
-            return new CommandResult(String.format(MESSAGE_FIND_NAME_PERSON_SUCCESS, predicate.getDisplayString()));
+            return new CommandResult(String.format(MESSAGE_FIND_EMAIL_PERSON_SUCCESS, predicate.getDisplayString()));
         } else {
             return new CommandResult(MESSAGE_FIND_PERSON_UNSUCCESSFUL);
         }
@@ -43,7 +43,7 @@ public class FindNameCommand extends FindCommand {
         }
 
         // instanceof handles nulls
-        if (!(other instanceof FindNameCommand otherFindCommand)) {
+        if (!(other instanceof FindEmailCommand otherFindCommand)) {
             return false;
         }
