+++ src/main/java/seedu/address/logic/commands/findcommand/FindPhoneCommand.java
@@ -1,7 +1,8 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.findcommand;
 
 import static java.util.Objects.requireNonNull;
 
+import seedu.address.logic.commands.CommandResult;
 import seedu.address.model.Model;
 import seedu.address.model.person.PhoneContainsKeywordsPredicate;
 /**
