+++ src/test/java/seedu/address/logic/parser/FindTagCommandParserTest.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.FindTagCommand;
+import seedu.address.logic.commands.findcommand.FindTagCommand;
 import seedu.address.model.person.TagContainsKeywordsPredicate;
 
 public class FindTagCommandParserTest {
